backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
openai==1.3.7
pypdfium2==4.26.0
python-docx==1.1.0
python-pptx==0.6.23
pytesseract==0.3.10
Pillow==10.1.0
google-cloud-vision==3.5.0
gunicorn==21.2.0
